requirements.txt
# VideoSort Pro v2 - Dependencias

# Interfaz gráfica (incluida en Python)
# tkinter

# Procesamiento de video e imágenes
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0

# Reconocimiento facial
face-recognition>=1.3.0
dlib>=19.22.0

# OCR (Reconocimiento óptico de caracteres)
pytesseract>=0.3.8

# APIs y requests HTTP
requests>=2.25.0

# Procesamiento de archivos
pathlib2>=2.3.0

# Progress bars y utilidades
tqdm>=4.60.0

# Logging mejorado
logging>=0.4.9.6

# JSON y configuración (incluido en Python)
# json

# Threading (incluido en Python)
# threading

# Regex (incluido en Python)
# re

# Sistema de archivos (incluido en Python)
# shutil
# os

# Fechas y tiempo (incluido en Python)
# datetime

# Tipos (incluido en Python 3.5+)
# typing
